test browser security and persistence contracts
