--- apps/desktop/src/main/services/builtInBrowser/builtInBrowserAgentAccess.test.ts
@@ -1,122 +1,0 @@
-import { describe, expect, it, vi } from "vitest";
-import type { Session } from "electron";
-import { createBuiltInBrowserAgentAccessController } from "./builtInBrowserAgentAccess";
-
-vi.mock("electron", () => ({
-  dialog: { showMessageBox: vi.fn() },
-}));
-
-function fakeSession(cookies: Array<{ domain: string }> = []): Session {
-  return {
-    cookies: {
-      get: vi.fn(async (filter: { url?: string }) => {
-        if (!filter.url) return cookies;
-        const host = new URL(filter.url).hostname;
-        return cookies.filter((cookie) => host === cookie.domain || host.endsWith(`.${cookie.domain}`));
-      }),
-    },
-  } as unknown as Session;
-}
-
-describe("builtInBrowserAgentAccess", () => {
-  it("allows unbound humans and local development origins without prompting", async () => {
-    const prompt = vi.fn(async () => ({ granted: false }));
-    const controller = createBuiltInBrowserAgentAccessController({
-      getSession: () => fakeSession(),
-      resolveParentWindow: () => null,
-      prompt,
-    });
-
-    await expect(controller.requireUrlAccess("https://github.com", {}, "test")).resolves.toBeUndefined();
-    await expect(controller.requireUrlAccess("http://localhost:5173", { chatSessionId: "chat-1" }, "test"))
-      .resolves.toBeUndefined();
-    expect(prompt).not.toHaveBeenCalled();
-  });
-
-  it("requires one chat-scoped human approval for high-risk origins", async () => {
-    const prompt = vi.fn(async () => ({ granted: true }));
-    const controller = createBuiltInBrowserAgentAccessController({
-      getSession: () => fakeSession(),
-      resolveParentWindow: () => null,
-      prompt,
-    });
-
-    await expect(controller.requireUrlAccess(
-      "https://github.com/settings/tokens",
-      { laneId: "lane-1", chatSessionId: "chat-1" },
-      "navigate",
-    )).resolves.toBeUndefined();
-    controller.assertUrlAccessSync("https://github.com/settings/tokens", { chatSessionId: "chat-1" });
-    expect(() => controller.assertUrlAccessSync(
-      "https://github.com/settings/tokens",
-      { chatSessionId: "chat-2" },
-    )).toThrow(/human-approval check/);
-    expect(prompt).toHaveBeenCalledTimes(1);
-  });
-
-  it("treats any origin with global-profile cookies as authenticated", async () => {
-    const session = fakeSession([{ domain: "example.com" }]);
-    const prompt = vi.fn(async () => ({ granted: false }));
-    const controller = createBuiltInBrowserAgentAccessController({
-      getSession: () => session,
-      resolveParentWindow: () => null,
-      prompt,
-    });
-
-    await expect(controller.requireUrlAccess(
-      "https://app.example.com/dashboard",
-      { chatSessionId: "chat-1" },
-      "observe",
-    )).rejects.toThrow(/Human approval was denied/);
-    expect(prompt).toHaveBeenCalledTimes(1);
-  });
-
-  it("requires first-use approval even when a non-local origin has no cookies", async () => {
-    const prompt = vi.fn(async () => ({ granted: true }));
-    const controller = createBuiltInBrowserAgentAccessController({
-      getSession: () => fakeSession(),
-      resolveParentWindow: () => null,
-      prompt,
-    });
-    const identity = { chatSessionId: "chat-1" };
-
-    expect(() => controller.assertUrlAccessSync("https://example.test", identity)).toThrow();
-    await expect(controller.requireUrlAccess("https://example.test", identity, "navigate")).resolves.toBeUndefined();
-    expect(() => controller.assertUrlAccessSync("https://example.test", identity)).not.toThrow();
-    expect(prompt).toHaveBeenCalledTimes(1);
-  });
-
-  it("requires approval for local origins with a remembered privileged permission", async () => {
-    const prompt = vi.fn(async () => ({ granted: false }));
-    const controller = createBuiltInBrowserAgentAccessController({
-      getSession: () => fakeSession(),
-      hasAllowedPermissionForOrigin: (origin) => origin === "http://localhost:5173",
-      resolveParentWindow: () => null,
-      prompt,
-    });
-
-    await expect(controller.requireUrlAccess(
-      "http://localhost:5173/account",
-      { chatSessionId: "chat-1" },
-      "navigate",
-    )).rejects.toThrow(/Human approval was denied/);
-    expect(prompt).toHaveBeenCalledTimes(1);
-  });
-
-  it("treats HTTP-authenticated origins as sensitive and grants only the prompting agent", async () => {
-    const controller = createBuiltInBrowserAgentAccessController({
-      getSession: () => fakeSession(),
-      resolveParentWindow: () => null,
-      prompt: vi.fn(),
-    });
-    const owner = { laneId: "lane-1", chatSessionId: "chat-1" };
-
-    controller.recordHumanAuthentication("https://basic.example.com/private", owner);
-
-    expect(() => controller.assertUrlAccessSync("https://basic.example.com/private", owner)).not.toThrow();
-    expect(() => controller.assertUrlAccessSync(
-      "https://basic.example.com/private",
-      { laneId: "lane-1", chatSessionId: "chat-2" },
-    )).toThrow(/requires a browser human-approval check/);
-  });
-});
