+++ docs/features/terminals-and-sessions/README.md
@@ -909,7 +909,11 @@ sibling, so `ade terminal read --chat-session <id>` always resolves a
 sensible target for attached-session agents. Every PTY launched through
 `ptyService.create` runs through `withAdeTerminalContextEnv` which
 exports `ADE_PROJECT_ROOT`, `ADE_LANE_ID`, and (when the PTY is
-session-owned) `ADE_CHAT_SESSION_ID` into the spawn env — that's how a
+session-owned) `ADE_CHAT_SESSION_ID` plus an opaque
+`ADE_BROWSER_ACTOR_TOKEN` into the spawn env. The browser capability is
+bound in memory to that owner chat/lane/project and lets the runtime reject
+an unbound shell or forged browser scope before it reaches Electron. The
+remaining identity variables are how a
 plain shell that the user types `ade --socket terminal read --chat-session
 "$ADE_CHAT_SESSION_ID" --text` into will resolve to the owning session's
 terminal even though no agent runtime spawned it. The headless ADE
